requirements.txt
Flask
python-dotenv
pymysql
SQLAlchemy
pandas
flask-restx
Flask-Cors
Flask-SQLAlchemy